Commit 21Oct
Added a confirmation window to ask user if they are sure they want to buy or sell. If not they can cancel out of it.

Added the functionality on stocks page that displays the total price of the stocks as user inputs quantity of shares to buy.

Added the functionality on portfolio page where as the user selects quantity of shares to sell, the total profit for the user displays.

Note: These functionality were originally done by Christian. Unfortunately Christian installed SQL Lite and it broke the app when we tried to pull. I had to re-do these functionalities and commit, but Christian should also get credit as he also did the work, just could not commit due to the database mishap.